requirements.txt
appdirs==1.4.3
appnope==0.1.0
decorator==4.0.11
Django==1.10.6
django-extensions==1.7.7
djangorestframework==3.6.2
djangorestframework-jwt==1.9.0
ipython==5.3.0
ipython-genutils==0.2.0
itsdangerous==0.24
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.2.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
